+++ libs/astarte_fdo/lib/ownership_voucher/ownership_voucher.ex
@@ -41,7 +41,9 @@ defmodule Astarte.FDO.OwnershipVoucher do
 
   The corresponding registration on the FDO rendezvous server is revoked first;
   if that fails, the voucher is not deleted.
+
   """
+  @spec delete(String.t(), binary()) :: :ok | {:error, term()}
   def delete(realm, guid) do
     with {:ok, ownership_voucher} <- Queries.fetch_ownership_voucher(guid),
          :ok <- ensure_voucher_in_realm(ownership_voucher.realm, realm),